+++ src/cratedb_fivetran_destination/cli.py
@@ -11,6 +11,7 @@ logger = logging.getLogger()
 @click.command()
 @click.version_option()
 @click.pass_context
+@click.option("--host", "-h", type=str, default="[::]", help="Host to listen on. Default: [::]")
 @click.option("--port", "-p", type=int, default=50052, help="Port to listen on. Default: 50052")
 @click.option(
     "--max-workers",
@@ -19,20 +20,18 @@ logger = logging.getLogger()
     default=1,
     help="The maximum number of threads that can be used. Default: 1",
 )
-def main(ctx: click.Context, port: int, max_workers: int) -> None:
+def main(ctx: click.Context, host: str, port: int, max_workers: int) -> None:
     """
     Start Fivetran CrateDB Destination gRPC server.
 
-    The executable needs to do the following:
+    Options:
 
-    - Accept a --port argument that takes an integer as a port number to listen to.
-    - Listen on both IPV4 (i.e. 0.0.0.0) and IPV6 (i.e ::0), but if only one is possible,
-      it should listen on IPV4.
+        --port: Port number to listen on. By default, listen on port 50052.
 
-    -- https://github.com/fivetran/fivetran_sdk/blob/main/development-guide.md#command-line-arguments
+        --host: Host to listen on. By default, listen on both IPV4 (0.0.0.0) and IPV6 (::0).
     """
     setup_logging()
-    server = start_server(port=port, max_workers=max_workers)
-    logger.info(f"Fivetran CrateDB Destination gRPC server started on port {port}")
+    server = start_server(host=host, port=port, max_workers=max_workers)
+    logger.info(f"Fivetran CrateDB Destination gRPC server started on {host}:{port}")
     server.wait_for_termination()
     logger.info("Fivetran CrateDB Destination gRPC server terminated")